backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
sqlmodel==0.0.22
httpx==0.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
